btrfs: add device_cache parameter to mount operations.
Allow btrfs to handle both store and cache devices. We add the device_cache
option to let the user explicit select what devices are used for caching. The
superblock for the device will in the future require superblock->dev_item->type
to be equal to 1 (BTRFS_DEV_ITEM_CACHE) to allow being mounted.