blk-mq: add new tagging scheme
Use the same idea as Kent for managing the free tags, keeping
them in an array and just keeping track of the free ones.
But augment that with being able to kick a prune of a local
cache, so we always have the full depth available. If not,
you could have old entries setting in now dormant CPUs that
never get reclaimed.

Also has support for reserved tags. Lots of drivers reserve
some tags (typically just 1) to be able to do error handling,
since error handling may require tagged IO as well.

Signed-off-by: Jens Axboe <axboe@kernel.dk>