blk-mq: add support for automatic rq payload data
Drivers always need to associate data with a request. Instead
of having them either waste time doing this dynamically in
the IO path, or by having to maintain and track this information
themselves, allow them to pass in the required per-request
data size at queue init time.

When a driver enables this and receives a request for IO,
the payload is stored in rq->special. In the current implementation,
the payload immediately follows the request in memory.

Signed-off-by: Jens Axboe <axboe@kernel.dk>